Detect merges where branch has been deleted, too
