--- scm-plugins/scm-git-plugin/src/main/java/sonia/scm/repository/api/GitPostReceiveHookMergeDetectionProvider.java
@@ -1,52 +1,0 @@
-/*
- * MIT License
- *
- * Copyright (c) 2020-present Cloudogu GmbH and Contributors
- *
- * Permission is hereby granted, free of charge, to any person obtaining a copy
- * of this software and associated documentation files (the "Software"), to deal
- * in the Software without restriction, including without limitation the rights
- * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
- * copies of the Software, and to permit persons to whom the Software is
- * furnished to do so, subject to the following conditions:
- *
- * The above copyright notice and this permission notice shall be included in all
- * copies or substantial portions of the Software.
- *
- * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
- * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
- * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
- * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
- * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
- * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
- * SOFTWARE.
- */
-
-package sonia.scm.repository.api;
-
-import org.eclipse.jgit.lib.Repository;
-import sonia.scm.repository.ChangesetPagingResult;
-import sonia.scm.repository.spi.GitLogComputer;
-import sonia.scm.repository.spi.HookMergeDetectionProvider;
-import sonia.scm.repository.spi.LogCommandRequest;
-
-public class GitPostReceiveHookMergeDetectionProvider implements HookMergeDetectionProvider {
-  private final Repository repository;
-  private final String repositoryId;
-
-  public GitPostReceiveHookMergeDetectionProvider(Repository repository, String repositoryId) {
-    this.repository = repository;
-    this.repositoryId = repositoryId;
-  }
-
-  @Override
-  public boolean branchesMerged(String target, String branch) {
-    LogCommandRequest request = new LogCommandRequest();
-    request.setBranch(branch);
-    request.setAncestorChangeset(target);
-    request.setPagingLimit(1);
-
-    ChangesetPagingResult changesets = new GitLogComputer(repositoryId, repository).compute(request);
-    return changesets.getTotal() == 0;
-  }
-}
