+++ scm-plugins/scm-git-plugin/src/main/java/sonia/scm/repository/api/GitReceiveHookMergeDetectionProvider.java
@@ -24,33 +24,58 @@
 
 package sonia.scm.repository.api;
 
+import org.eclipse.jgit.lib.AnyObjectId;
+import org.eclipse.jgit.lib.ObjectId;
 import org.eclipse.jgit.lib.Repository;
+import org.eclipse.jgit.transport.ReceiveCommand;
+import sonia.scm.repository.GitUtil;
 import sonia.scm.repository.spi.GitLogComputer;
 import sonia.scm.repository.spi.HookMergeDetectionProvider;
 import sonia.scm.repository.spi.LogCommandRequest;
 
+import java.util.List;
+
 public class GitReceiveHookMergeDetectionProvider implements HookMergeDetectionProvider {
   private final Repository repository;
   private final String repositoryId;
-  private final BranchMapper branchMapper;
+  private final List<ReceiveCommand> receiveCommands;
 
-  public GitReceiveHookMergeDetectionProvider(Repository repository, String repositoryId, BranchMapper branchMapper) {
+  public GitReceiveHookMergeDetectionProvider(Repository repository, String repositoryId, List<ReceiveCommand> receiveCommands) {
     this.repository = repository;
     this.repositoryId = repositoryId;
-    this.branchMapper = branchMapper;
+    this.receiveCommands = receiveCommands;
   }
 
   @Override
   public boolean branchesMerged(String target, String branch) {
     LogCommandRequest request = new LogCommandRequest();
-    request.setBranch(branchMapper.map(branch));
-    request.setAncestorChangeset(branchMapper.map(target));
+    request.setBranch(findRelevantRevisionForBranchIfToBeUpdated(branch));
+    request.setAncestorChangeset(findRelevantRevisionForBranchIfToBeUpdated(target));
     request.setPagingLimit(1);
 
     return new GitLogComputer(repositoryId, repository).compute(request).getTotal() == 0;
   }
 
-  public interface BranchMapper {
-    String map(String branch);
+  private String findRelevantRevisionForBranchIfToBeUpdated(String branch) {
+    return receiveCommands
+      .stream()
+      .filter(receiveCommand -> isReceiveCommandForBranch(branch, receiveCommand))
+      .map(this::getRelevantRevision)
+      .map(AnyObjectId::getName)
+      .findFirst()
+      .orElse(branch);
+  }
+
+  private boolean isReceiveCommandForBranch(String branch, ReceiveCommand receiveCommand) {
+    return receiveCommand.getType() != ReceiveCommand.Type.CREATE
+      && GitUtil.getBranch(receiveCommand.getRef()).equals(branch);
+  }
+
+  private ObjectId getRelevantRevision(ReceiveCommand receiveCommand) {
+    if (receiveCommand.getType() == ReceiveCommand.Type.DELETE) {
+      return receiveCommand.getOldId();
+    } else {
+      return receiveCommand.getNewId();
+    }
   }
 }
